ft: manage extensions configuration
Each extension is now responsible to manage its config stored in the
global config.json, through a custom configValidator function.
