+++ conf/Config.js
@@ -4,8 +4,8 @@ const fs = require('fs');
 const path = require('path');
 const joi = require('joi');
 
+const extensions = require('../extensions');
 const backbeatConfigJoi = require('./config.joi.js');
-const { adminCredsJoi } = require('../lib/config/configItems.joi.js');
 
 class Config {
     constructor() {
@@ -15,15 +15,15 @@ class Config {
          */
         this._basePath = __dirname;
         if (process.env.BACKBEAT_CONFIG_FILE !== undefined) {
-            this.configPath = process.env.BACKBEAT_CONFIG_FILE;
+            this._configPath = process.env.BACKBEAT_CONFIG_FILE;
         } else {
-            this.configPath = path.join(this._basePath, 'config.json');
+            this._configPath = path.join(this._basePath, 'config.json');
         }
 
         let config;
         try {
-            const data = fs.readFileSync(this.configPath,
-              { encoding: 'utf-8' });
+            const data = fs.readFileSync(this._configPath,
+                                         { encoding: 'utf-8' });
             config = JSON.parse(data);
         } catch (err) {
             throw new Error(`could not parse config file: ${err.message}`);
@@ -32,61 +32,31 @@ class Config {
         const parsedConfig = joi.attempt(config, backbeatConfigJoi,
                                          'invalid backbeat config');
 
+        if (parsedConfig.extensions) {
+            Object.keys(parsedConfig.extensions).forEach(extName => {
+                const index = extensions[extName];
+                if (!index) {
+                    throw new Error(`configured extension ${extName}: ` +
+                                    'not found in extensions directory');
+                }
+                if (index.configValidator) {
+                    const extConfig = parsedConfig.extensions[extName];
+                    const validatedConfig =
+                              index.configValidator(this, extConfig);
+                    parsedConfig.extensions[extName] = validatedConfig;
+                }
+            });
+        }
         // config is validated, safe to assign directly to the config object
         Object.assign(this, parsedConfig);
-
-        if (this.extensions !== undefined &&
-            this.extensions.replication !== undefined) {
-            const { source, destination } = this.extensions.replication;
-
-            if (source.auth.vault) {
-                const { adminCredentialsFile } = source.auth.vault;
-                if (adminCredentialsFile) {
-                    source.auth.vault.adminCredentials =
-                        this._loadAdminCredentialsFromFile(
-                            adminCredentialsFile);
-                }
-            }
-            if (destination.auth.vault) {
-                const { adminCredentialsFile } = destination.auth.vault;
-                if (adminCredentialsFile) {
-                    destination.auth.vault.adminCredentials =
-                        this._loadAdminCredentialsFromFile(
-                            adminCredentialsFile);
-                }
-            }
-
-            // additional target certs checks
-            const { key, cert, ca } = destination.certFilePaths;
-
-            const makePath = value => (value.startsWith('/') ?
-                                       value : `${this._basePath}/${value}`);
-            const keypath = makePath(key);
-            const certpath = makePath(cert);
-            let capath = undefined;
-            fs.accessSync(keypath, fs.F_OK | fs.R_OK);
-            fs.accessSync(certpath, fs.F_OK | fs.R_OK);
-            if (ca) {
-                capath = makePath(ca);
-                fs.accessSync(capath, fs.F_OK | fs.R_OK);
-            }
-
-            this.extensions.replication.destination.https = {
-                cert: fs.readFileSync(certpath, 'ascii'),
-                key: fs.readFileSync(keypath, 'ascii'),
-                ca: ca ? fs.readFileSync(capath, 'ascii') : undefined,
-            };
-        }
     }
 
-    _loadAdminCredentialsFromFile(filePath) {
-        const adminCredsJSON = fs.readFileSync(filePath);
-        const adminCredsObj = JSON.parse(adminCredsJSON);
-        joi.attempt(adminCredsObj, adminCredsJoi,
-                    'invalid admin credentials');
-        const accessKey = Object.keys(adminCredsObj)[0];
-        const secretKey = adminCredsObj[accessKey];
-        return { accessKey, secretKey };
+    getBasePath() {
+        return this._basePath;
+    }
+
+    getConfigPath() {
+        return this._configPath;
     }
 }
