+++ conf/config.joi.js
@@ -1,11 +1,7 @@
 'use strict'; // eslint-disable-line
 
 const joi = require('joi');
-const { hostPortJoi, bootstrapListJoi, logJoi } =
-    require('../lib/config/configItems.joi.js');
-
-const transportJoi = joi.alternatives().try('http', 'https')
-    .default('http');
+const { hostPortJoi, logJoi } = require('../lib/config/configItems.joi.js');
 
 const joiSchema = {
     zookeeper: {
@@ -27,63 +23,7 @@ const joiSchema = {
         }).when('logSource', { is: 'dmd', then: joi.required() }),
     },
     log: logJoi,
-    extensions: {
-        replication: {
-            source: {
-                transport: transportJoi,
-                s3: hostPortJoi.required(),
-                auth: joi.object({
-                    type: joi.alternatives().try('account', 'role').required(),
-                    account: joi.string()
-                        .when('type', { is: 'account', then: joi.required() }),
-                    vault: joi.object({
-                        host: joi.string().required(),
-                        port: joi.number().greater(0).required(),
-                        adminPort: joi.number().greater(0)
-                            .when('adminCredentialsFile', {
-                                is: joi.exist(),
-                                then: joi.required(),
-                            }),
-                        adminCredentialsFile: joi.string().optional(),
-                    }).when('type', { is: 'role', then: joi.required() }),
-                }).required(),
-            },
-            destination: {
-                transport: transportJoi,
-                auth: joi.object({
-                    type: joi.alternatives().try('account', 'role').required(),
-                    account: joi.string()
-                        .when('type', { is: 'account', then: joi.required() }),
-                    vault: joi.object({
-                        host: joi.string().optional(),
-                        port: joi.number().greater(0).optional(),
-                        adminPort: joi.number().greater(0).optional(),
-                        adminCredentialsFile: joi.string().optional(),
-                    }),
-                }).required(),
-                bootstrapList: bootstrapListJoi,
-                certFilePaths: joi.object({
-                    key: joi.string().required(),
-                    cert: joi.string().required(),
-                    ca: joi.string().empty(''),
-                }).required(),
-            },
-            topic: joi.string().required(),
-            replicationStatusTopic: joi.string().required(),
-            queueProcessor: {
-                groupId: joi.string().required(),
-                retryTimeoutS: joi.number().default(300),
-                // versioning can support out of order updates
-                concurrency: joi.number().greater(0).default(10),
-            },
-            replicationStatusProcessor: {
-                groupId: joi.string().required(),
-                retryTimeoutS: joi.number().default(300),
-                // versioning can support out of order updates
-                concurrency: joi.number().greater(0).default(10),
-            },
-        },
-    },
+    extensions: joi.object(),
 };
 
 module.exports = joiSchema;
