+++ extensions/replication/index.js
@@ -1,7 +1,9 @@
+const ReplicationConfigValidator = require('./ReplicationConfigValidator');
 const ReplicationQueuePopulator = require('./ReplicationQueuePopulator');
 
 module.exports = {
     name: 'replication',
     version: '1.0.0',
+    configValidator: ReplicationConfigValidator,
     queuePopulatorExtension: ReplicationQueuePopulator,
 };
