+++ lib/queuePopulator/QueuePopulator.js
@@ -152,13 +152,15 @@ class QueuePopulator {
         Object.keys(this.extConfigs).forEach(extName => {
             const extConfig = this.extConfigs[extName];
             const index = require(`../../extensions/${extName}/index.js`);
-            // eslint-disable-next-line new-cap
-            const ext = new index.queuePopulatorExtension({
-                config: extConfig,
-                logger: this.log,
-            });
-            this.log.info(`${index.name} extension is active`);
-            this._extensions.push(ext);
+            if (index.queuePopulatorExtension) {
+                // eslint-disable-next-line new-cap
+                const ext = new index.queuePopulatorExtension({
+                    config: extConfig,
+                    logger: this.log,
+                });
+                this.log.info(`${index.name} extension is active`);
+                this._extensions.push(ext);
+            }
         });
     }
